requirements.txt
streamlit
python-dotenv
groq
langchain-community
python-docx
PyPDF2
sentence-transformers
faiss-cpu
pandas
